requirements.txt
fastapi
uvicorn
numpy
pillow
torch
torchvision
tqdm
opencv-python
scikit-image
matplotlib
addict
dlib-bin
joblib
scikit-learn
imageio
lpips
wandb
typing_extensions
typing-inspection
gdown
dill
face-alignment
filelock
regex
requests
protobuf
pydantic
setproctitle
six
python-multipart
git+https://github.com/openai/CLIP.git
